Do not attempt to disconnect AD if we failed to install MSOnline

+++ deployment/safe_haven_management_environment/remote/create_dc/artifacts/shm-dc1-configuration/Disconnect_AD.template.ps1
@@ -6,13 +6,15 @@ if (-Not (Get-Module -ListAvailable -Name MSOnline)) {
     Install-Module -Name MSOnline -Force
 }
 
-Write-Output "Please use username admin@<shm-fqdn> and the <aad-admin-password-name> from <shm-keyvault-name>."
-Connect-MsolService
-Write-Output "Is directory synchronisation currently enabled? $((Get-MSOLCompanyInformation).DirectorySynchronizationEnabled)"
-Write-Output "Disabling directory synchronisation..."
-Set-MsolDirSyncEnabled -EnableDirSync $False -Force
-Write-Output "Is directory synchronisation currently enabled? $((Get-MSOLCompanyInformation).DirectorySynchronizationEnabled)"
-# Remove user-added service principals except the MFA service principal
-Write-Output "Removing any user-added service principals..."
-Get-MsolServicePrincipal | Where-Object { $_.AppPrincipalId -ne "981f26a1-7f43-403b-a875-f8b09b8cd720" } | Remove-MsolServicePrincipal 2>&1 | Out-Null
-Write-Output "Finished"
+if (Get-Module -ListAvailable -Name MSOnline) {
+    Write-Output "Please use username admin@<shm-fqdn> and the <aad-admin-password-name> from <shm-keyvault-name>."
+    Connect-MsolService
+    Write-Output "Is directory synchronisation currently enabled? $((Get-MSOLCompanyInformation).DirectorySynchronizationEnabled)"
+    Write-Output "Disabling directory synchronisation..."
+    Set-MsolDirSyncEnabled -EnableDirSync $False -Force
+    Write-Output "Is directory synchronisation currently enabled? $((Get-MSOLCompanyInformation).DirectorySynchronizationEnabled)"
+    # Remove user-added service principals except the MFA service principal
+    Write-Output "Removing any user-added service principals..."
+    Get-MsolServicePrincipal | Where-Object { $_.AppPrincipalId -ne "981f26a1-7f43-403b-a875-f8b09b8cd720" } | Remove-MsolServicePrincipal 2>&1 | Out-Null
+    Write-Output "Finished"
+}
